+++ node/src/RedisClient.ts
@@ -102,23 +102,14 @@ export class RedisClient extends BaseClient {
      *      If the transaction failed due to a WATCH command, `exec` will return `null`.
      */
     public exec(transaction: Transaction): Promise<ReturnType[] | null> {
-        return this.createWritePromise(transaction.commands).then(
-            (result: unknown) => {
-                if (result === null) {
-                    return null;
-                }
-
-                const modifiedResult: ReturnType[] = result as ReturnType[];
-
-                for (const index of transaction.set_commands_indexes) {
-                    modifiedResult[index] = new Set(
-                        modifiedResult[index] as string[],
-                    );
-                }
-
-                return modifiedResult;
-            },
-        );
+        return this.createWritePromise<ReturnType[] | null>(
+            transaction.commands,
+        ).then((result: ReturnType[] | null) => {
+            return this.processResultWithSetCommands(
+                result,
+                transaction.set_commands_indexes,
+            );
+        });
     }
 
     /** Executes a single command, without checking inputs. Every part of the command, including subcommands,